Added 401 to /logout if authorization token not provided

+++ controllers/AuthController.js
@@ -28,9 +28,11 @@ module.exports = {
 
 	logout: async (req,res,next) => {
 		try {
-			const { token_id } = jwt.verify(req.headers.authorization, jwt_secret);
-			const result = await Token.destroy({ where: { id: token_id }});
-			if(result > 0) return res.send({ message: 'Logged out' });
+			if(req.headers.authorization) {
+				const { token_id } = jwt.verify(req.headers.authorization, jwt_secret);
+				const result = await Token.destroy({ where: { id: token_id }});
+				if(result > 0) return res.send({ message: 'Logged out' });
+			}
 			return res.status(401).send({ message: 'Unauthorized' });
 		} catch(error) {
 			console.log(error);
